refactor: move helpers to "coverage.ts"
